Rename stagingdir to staging_dir.
